Move (Hashed)CertificateValue to certificate/value.rs
