+++ linera-chain/src/certificate/confirmed.rs
@@ -5,10 +5,13 @@
 use linera_base::identifiers::{BlobId, ChainId, MessageId};
 use linera_execution::committee::Epoch;
 
-use super::{generic::GenericCertificate, hashed::Hashed, Certificate};
+use super::{
+    generic::GenericCertificate, hashed::Hashed, Certificate, CertificateValue,
+    HashedCertificateValue,
+};
 use crate::{
     block::{ConfirmedBlock, ValidatedBlock},
-    data_types::{CertificateValue, ExecutedBlock, HashedCertificateValue, Medium, MessageBundle},
+    data_types::{ExecutedBlock, Medium, MessageBundle},
 };
 
 impl GenericCertificate<ConfirmedBlock> {
