+++ linera-core/src/client/mod.rs
@@ -40,11 +40,13 @@ use linera_base::{
 };
 use linera_chain::{
     data_types::{
-        Block, BlockProposal, ChainAndHeight, ExecutedBlock, HashedCertificateValue,
-        IncomingBundle, LiteCertificate, LiteVote, MessageAction,
+        Block, BlockProposal, ChainAndHeight, ExecutedBlock, IncomingBundle, LiteCertificate,
+        LiteVote, MessageAction,
     },
     manager::ChainManagerInfo,
-    types::{Certificate, ConfirmedBlockCertificate, ValidatedBlockCertificate},
+    types::{
+        Certificate, ConfirmedBlockCertificate, HashedCertificateValue, ValidatedBlockCertificate,
+    },
     ChainError, ChainExecutionContext, ChainStateView,
 };
 use linera_execution::{
