+++ linera-core/src/unit_tests/worker_tests.rs
@@ -28,11 +28,11 @@ use linera_base::{
 use linera_chain::{
     data_types::{
         Block, BlockExecutionOutcome, BlockProposal, ChainAndHeight, ChannelFullName,
-        HashedCertificateValue, IncomingBundle, LiteVote, Medium, MessageAction, MessageBundle,
-        Origin, OutgoingMessage, PostedMessage, SignatureAggregator,
+        IncomingBundle, LiteVote, Medium, MessageAction, MessageBundle, Origin, OutgoingMessage,
+        PostedMessage, SignatureAggregator,
     },
     test::{make_child_block, make_first_block, BlockTestExt, MessageTestExt, VoteTestExt},
-    types::Certificate,
+    types::{Certificate, HashedCertificateValue},
     ChainError, ChainExecutionContext,
 };
 use linera_execution::{
