+++ linera-indexer/lib/src/plugin.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 
 use async_graphql::{EmptyMutation, EmptySubscription, ObjectType, Schema};
 use axum::Router;
-use linera_chain::data_types::HashedCertificateValue;
+use linera_chain::types::HashedCertificateValue;
 use linera_views::{context::ViewContext, store::KeyValueStore, views::View};
 use tokio::sync::Mutex;
