+++ linera-rpc/src/client.rs
@@ -7,8 +7,8 @@ use linera_base::{
     identifiers::{BlobId, ChainId},
 };
 use linera_chain::{
-    data_types::{BlockProposal, HashedCertificateValue, LiteCertificate},
-    types::Certificate,
+    data_types::{BlockProposal, LiteCertificate},
+    types::{Certificate, HashedCertificateValue},
 };
 use linera_core::{
     data_types::{ChainInfoQuery, ChainInfoResponse},
