+++ linera-rpc/src/grpc/conversions.rs
@@ -8,11 +8,8 @@ use linera_base::{
     identifiers::{BlobId, ChainId, Owner},
 };
 use linera_chain::{
-    data_types::{
-        BlockProposal, CertificateValue, HashedCertificateValue, LiteCertificate, LiteValue,
-        ProposalContent,
-    },
-    types::Certificate,
+    data_types::{BlockProposal, LiteCertificate, LiteValue, ProposalContent},
+    types::{Certificate, CertificateValue, HashedCertificateValue},
 };
 use linera_core::{
     data_types::{ChainInfoQuery, ChainInfoResponse, CrossChainRequest},
@@ -712,8 +709,9 @@ pub mod tests {
         data_types::{Amount, Round, Timestamp},
     };
     use linera_chain::{
-        data_types::{Block, BlockExecutionOutcome, HashedCertificateValue},
+        data_types::{Block, BlockExecutionOutcome},
         test::make_first_block,
+        types::HashedCertificateValue,
     };
     use linera_core::data_types::ChainInfo;
     use serde::{Deserialize, Serialize};
